+++ designsystem/src/main/java/com/mundcode/designsystem/components/dialogs/ContentDialog.kt
@@ -92,45 +92,3 @@ fun SelectableDialog(
         }
     }
 }
-
-@Composable
-fun TotalNumberPickerDialog(
-    onResult: (Int) -> Unit,
-    onCancel: () -> Unit
-) {
-    var selectedNumber by remember {
-        mutableStateOf(1)
-    }
-
-    SelectableDialog(
-        title = "총 문항수 설정",
-        onClickClose = onCancel,
-        onClickCancel = onCancel,
-        onClickConfirm = {
-            onResult(selectedNumber)
-        }
-    ) {
-        LazyVerticalGrid(
-            columns = GridCells.Fixed(7),
-            modifier = Modifier
-                .height(280.dp)
-                .fillMaxWidth()
-        ) {
-            items((1..200).toList()) { number ->
-                SelectableNumberText(
-                    number = number,
-                    isSelected = number == selectedNumber,
-                    onClickNumber = { selectedNumber = it })
-            }
-        }
-    }
-}
-
-@Preview(backgroundColor = 0xFFFFFF)
-@Composable
-fun TotalNumberPickerDialogPreview() {
-    TotalNumberPickerDialog(
-        onResult = {},
-        onCancel = {}
-    )
-}
